+++ src/task.rs
@@ -1,51 +1,51 @@
 use crate::task_list::TaskList;
 use core::{
     ops::Range,
-    sync::atomic::{AtomicUsize, Ordering},
+    sync::atomic::{AtomicU64, Ordering},
 };
 
 #[derive(Debug)]
 pub struct Task {
-    pub(crate) start: AtomicUsize,
-    pub(crate) end: AtomicUsize,
+    pub(crate) start: AtomicU64,
+    pub(crate) end: AtomicU64,
 }
 
 impl Task {
-    pub fn remain(&self) -> usize {
+    pub fn remain(&self) -> u64 {
         let start = self.start();
         let end = self.end();
         end.checked_sub(start).unwrap_or(0)
     }
 
-    pub fn start(&self) -> usize {
+    pub fn start(&self) -> u64 {
         self.start.load(Ordering::Acquire)
     }
-    pub fn set_start(&self, start: usize) {
+    pub fn set_start(&self, start: u64) {
         self.start.store(start, Ordering::Release);
     }
-    pub fn fetch_add_start(&self, value: usize) -> usize {
+    pub fn fetch_add_start(&self, value: u64) -> u64 {
         self.start.fetch_add(value, Ordering::AcqRel)
     }
-    pub fn fetch_sub_start(&self, value: usize) -> usize {
+    pub fn fetch_sub_start(&self, value: u64) -> u64 {
         self.start.fetch_sub(value, Ordering::AcqRel)
     }
-    pub fn end(&self) -> usize {
+    pub fn end(&self) -> u64 {
         self.end.load(Ordering::Acquire)
     }
-    pub fn set_end(&self, end: usize) {
+    pub fn set_end(&self, end: u64) {
         self.end.store(end, Ordering::Release);
     }
-    pub fn fetch_add_end(&self, value: usize) -> usize {
+    pub fn fetch_add_end(&self, value: u64) -> u64 {
         self.end.fetch_add(value, Ordering::AcqRel)
     }
-    pub fn fetch_sub_end(&self, value: usize) -> usize {
+    pub fn fetch_sub_end(&self, value: u64) -> u64 {
         self.end.fetch_sub(value, Ordering::AcqRel)
     }
 
-    pub fn new(start: usize, end: usize) -> Self {
+    pub fn new(start: u64, end: u64) -> Self {
         Self {
-            start: AtomicUsize::new(start),
-            end: AtomicUsize::new(end),
+            start: AtomicU64::new(start),
+            end: AtomicU64::new(end),
         }
     }
 }
@@ -56,14 +56,14 @@ impl PartialEq for Task {
     }
 }
 
-impl From<&(usize, usize)> for Task {
-    fn from(value: &(usize, usize)) -> Self {
+impl From<&(u64, u64)> for Task {
+    fn from(value: &(u64, u64)) -> Self {
         Self::new(value.0, value.1)
     }
 }
 
-impl From<&Range<usize>> for Task {
-    fn from(value: &Range<usize>) -> Self {
+impl From<&Range<u64>> for Task {
+    fn from(value: &Range<u64>) -> Self {
         Self::new(value.start, value.end)
     }
 }
